requirements.txt
blinker==1.9.0
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
Flask==3.1.1
fonttools==4.58.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pymongo==4.13.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
seaborn==0.13.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
